% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medline.R
\name{read_medline}
\alias{read_medline}
\title{Read Pubmed Format}
\usage{
read_medline(text, key1 = "PMID-", key2 = "AB  -")
}
\arguments{
\item{text}{text plain (oubmed format) of journal articles}

\item{key1}{"PMID-" to obtain the PMID of each abstract}

\item{key2}{"AB  -" to obtain the abstract of each PMID}
}
\value{
A data.frame with the PMID and Abstract
}
\description{
A function to read pubmed (MEDLINE) format text plane, to convert into data.frame
}
\examples{
example1 <- readLines("pubmed-Oligodendr_ODP-set.txt")
prueba <- read_medline(example1)

}
